split response stuff out into it's on class; cleaned up the hooks
